% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{blockdiag}
\alias{blockdiag}
\title{Block diagonal matrix}
\usage{
blockdiag(...)
}
\arguments{
\item{...}{matrices to construct a block diagonal matrix, can be a list also}
}
\value{
a block diagonal matrix
}
\description{
Block diagonal matrix
}
