% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{bunch}
\alias{bunch}
\title{Grouping the left hand side in %=%}
\usage{
bunch(...)
}
\arguments{
\item{...}{variables to be bunched}
}
\value{
a list of variable names
}
\description{
bunch multiple items together for %=%
}
