% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{colorise}
\alias{colorise}
\title{assign colours to distinct values of a vector}
\usage{
colorise(v)
}
\arguments{
\item{v}{vector to be converted to colours}
}
\value{
a vector of colours of equal length with two attributes: 'palette' giving the palette used for all distinct colours; 'uniq': vector of distinct values in v
}
\description{
assign colours to distinct values of a vector
}
