% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{CvM.test}
\alias{CvM.test}
\title{Cramer--von-Mise two sample test}
\usage{
CvM.test(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
a list containing test statistic and significance code
}
\description{
Cramer--von-Mise two sample test
}
