% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{dp}
\alias{dp}
\title{display decimal places nicely}
\usage{
dp(x, digits)
}
\arguments{
\item{x}{a real number}

\item{digits}{number of decimal places to keep}
}
\value{
a string
}
\description{
display decimal places nicely
}
\details{
keep trailing zeros
}
