% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{EM}
\alias{EM}
\title{EM algorithm}
\usage{
EM(X, num_clusters, tol = 1e-10)
}
\arguments{
\item{X}{data matrix of dimension n x p}

\item{num_clusters}{number of clusters}

\item{tol}{tolerance for convergence}
}
\value{
a list: mu_hat - the estimated cluster centroids and soft_label -
the estimated cluster membership
}
\description{
EM algorithm with identity covariance matrix for all groups
}
