% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate multiple expressions}
\usage{
evaluate(...)
}
\arguments{
\item{...}{expressions of the form name=val}
}
\value{
a list of variable names
}
\description{
useful to set all arguments in a function to their default values
}
