% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{expit}
\alias{expit}
\title{Inverse of the logistic function}
\usage{
expit(a)
}
\arguments{
\item{a}{a real number}
}
\value{
a real number of value 1/(exp(-a)+1)
}
\description{
Inverse of the logistic function
}
\seealso{
\code{\link{logit}}
}
