% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{filter_data_frame}
\alias{filter_data_frame}
\title{Filter data frame by condition on selected columns}
\usage{
filter_data_frame(df, cols, vals)
}
\arguments{
\item{df}{data frame}

\item{cols}{column indices or names to use for selection}

\item{vals}{a matrix whose rows consist of allowed parameter combinations}
}
\description{
return a subset of rows of the original dataframe where the
corresponding columns have values belonging to one of the rows of the vals
matrix
}
