% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{find.last}
\alias{find.last}
\title{Find the location of final TRUE value in a boolean vector}
\usage{
find.last(v)
}
\arguments{
\item{v}{a logical vector}
}
\value{
an integer denotating the location, return NA if not found.
}
\description{
Find the location of final TRUE value in a boolean vector
}
