% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{\%setminus\%}
\alias{\%setminus\%}
\title{Set minus: remove elements of small set from large set}
\usage{
large \%setminus\% small
}
\arguments{
\item{large}{the large set}

\item{small}{the small set}
}
\value{
large set with elements of small set removed
}
\description{
Set minus: remove elements of small set from large set
}
