% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdp.R
\name{init_project}
\alias{init_project}
\title{'Quickly Deployable Project (qdp)'}
\usage{
init_project(project, path = "", cores = NULL, modules = NULL,
  python = 3, ...)
}
\arguments{
\item{project}{The name of the project}

\item{path}{Path to project; default is home/user/project/the_name_of_project}

\item{cores}{The number of cores to initialize. Default is NULL which 
uses all available cores}
}
\description{
Creates a standardized scaffolding to quickly deploy an analysis(project)
}
\details{
'qdp' creates a scaffolding to enable an analyst to do 
what an analyst does best: analysis

Once a 'project' is created using 'create_project', the analyst may load
accompanying scripts, configuration files, data, reports, etc by running
'start_project'

The project's directory structure is specified as such:
project              # name of project
 │ 
 ├── README.md       # readme
 ├── .git/           # git repo 
 ├── inst/           # installation
 ├── data/           # data
 ├── lib/            # library code
 │   ├── class.py    # source code (e.g., .R, .py files)
 │   └── analysis.R  # analysis script
 ├── conf/
 │    ├── config.py  # python config files
 │    └── config.R   # R config files
 ├── lit/            # literature
 ├── report/         # reports
 └── log/            # log files

Optional arguments to override defaults: 

'r_config_options'       .Rprofile-like options; however, project specific

'py_config_options'      creates config.py which may imported as 
                         'import config' for python

'spark_config_options'   creates the default configuration for Spark


'module_setup_options'   other R scripts to source upon setup; 
                         e.g., source('project/lib/analysis.R')

'git_config_options'     git configuration options upon initialization

'git_ignore_options'     git ignore options upon initialization

The project's default path is set to: 

Linux:       /home/user/projects
Windows:     C:/Users/user/projects
Mac:         /Users/user/projects
}
\author{
Nathan Matare <email: nmatare@chicagobooth.com>
}
