% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{isString}
\alias{isString}
\title{Check whether input is string}
\usage{
isString(x)
}
\arguments{
\item{x}{object}
}
\value{
boolean for whether x is string
}
\description{
Check whether input is string
}
