% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{lambertW}
\alias{lambertW}
\title{The Lambert W function}
\usage{
lambertW(z, b = 0, maxiter = 10, eps = .Machine$double.eps,
  min.imag = 1e-09)
}
\arguments{
\item{z}{a complex number}

\item{b}{an integer, determining the branch of the W function}

\item{maxiter}{maximum number of Halley iteration}

\item{eps}{floating point tolerance}

\item{min.imag}{floating point tolerance for imaginary part}
}
\value{
a real number of value log(a/(1-a))
}
\description{
the Lambert W function is the inverse of f(x) = x*exp(x). The two main branches are b = 0 and b = -1.
}
