% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{logit}
\alias{logit}
\title{The logistic function}
\usage{
logit(a)
}
\arguments{
\item{a}{a real number}
}
\value{
a real number of value log(a/(1-a))
}
\description{
The logistic function
}
\seealso{
\code{\link{expit}}
}
