% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{match.data.frame}
\alias{match.data.frame}
\title{match rows of dataframes}
\usage{
match.data.frame(
  match_from,
  match_to,
  nomatch = NA_integer_,
  method = "string"
)
}
\arguments{
\item{match_from}{data frame to match from}

\item{match_to}{data frame to match to}

\item{nomatch}{what to do if there is no match}

\item{method}{choose between 'string' and 'list', the former compares entries by their string values (i.e. 10 and '10' are considered equal), the latter requires exact type match}
}
\description{
`match.data.frame` returns a vector of the positions of (first) matches of its first argument in its second. `%in%` is a more intuitive interface as a binary operator, which returns a logical vector indicating if there is a match or not for its left operand.
}
