% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{matplotlib_palette}
\alias{matplotlib_palette}
\title{matplotlib_palette
return the first n palette colours}
\usage{
matplotlib_palette(n = 0, scheme = "default", visualise = FALSE)
}
\arguments{
\item{n}{number of colours}

\item{scheme}{palette scheme, one of 'default', 'bright' and 'rainbow'}

\item{visualise}{if TRUE, a barplot of all colours will be shown}
}
\value{
a vector of hexadecimal colours
}
\description{
matplotlib_palette
return the first n palette colours
}
