% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{matrix.GramSchmidt}
\alias{matrix.GramSchmidt}
\title{Apply Gram Schmidt operation to a matrix}
\usage{
matrix.GramSchmidt(A, tolerance = 1e-14)
}
\arguments{
\item{A}{a matrix of real values}

\item{tolerance}{tolerance level for 0}
}
\value{
a matrix of the same size with orthonormal columns
}
\description{
Apply Gram Schmidt operation to a matrix
}
\details{
the resulting matrix is uniquely identified by asserting the diagonal to be non-negative
}
