% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{matrix.rank}
\alias{matrix.rank}
\title{Rank of a matrix}
\usage{
matrix.rank(A, tolerance = 1e-10)
}
\arguments{
\item{A}{a matrix of real numbers}

\item{tolerance}{level below which we treat a singular value to be 0}
}
\value{
an nonnegative integer
}
\description{
Rank of a matrix
}
