% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{matrix.standardise}
\alias{matrix.standardise}
\title{Standardise the columns of a matrix to have mean zero and length 1}
\usage{
matrix.standardise(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
a standardised matrix
}
\description{
Standardise the columns of a matrix to have mean zero and length 1
}
