% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{matrix.trace}
\alias{matrix.trace}
\title{Trace of a square matrix}
\usage{
matrix.trace(M)
}
\arguments{
\item{M}{a square matrix of real numbers}
}
\value{
a real number
}
\description{
Trace of a square matrix
}
