% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{mix_color}
\alias{mix_color}
\title{Color mixing}
\usage{
mix_color(color1, color2, lambda = 0.5)
}
\arguments{
\item{color1}{color 1 in hexadecimal}

\item{color2}{color 2 in hexadecimal}

\item{lambda}{interpolation, if equal to 0, return color1, if 1 return color2}
}
\description{
Color mixing
}
