% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{myplot}
\alias{myplot}
\title{generate a plot from a data frame with color and line type set by specific columns}
\usage{
myplot(
  x,
  y,
  col = NULL,
  style = NULL,
  data = .GlobalEnv,
  legend.position = "topright",
  ...
)
}
\arguments{
\item{x}{column name for the x values, needs to be a string}

\item{y}{column name for the y values, needs to be a string}

\item{col}{column name for the colour attributes, can be of any type, distinct values are represented by distinct colours}

\item{style}{column name for the line type attributes, can be of any type, distinct values are represented by distinct line types}

\item{data}{data frame}

\item{legend.position}{where to place the legend}

\item{...}{other plotting parameters}
}
\description{
generate a plot from a data frame with color and line type set by specific columns
}
