% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{offdiag}
\alias{offdiag}
\title{extracting the off-diagonal part of A}
\usage{
offdiag(A, as.vector = FALSE)
}
\arguments{
\item{A}{a rectangular matrix}

\item{as.vector}{if FALSE, a matrix with main diagonal set to 0 is returned
otherwise, a vector of offdiagonal entries is returned}
}
\description{
extracting the off-diagonal part of A
}
