% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{powerMethod}
\alias{powerMethod}
\title{Power method for leading eigenvectors and eigenvalues}
\usage{
powerMethod(A, k = 1, eps = 1e-10, maxiter = 1000)
}
\arguments{
\item{A}{a square matrix}

\item{k}{number of leading eigenvectors/eigenvalues}

\item{eps}{tolerance for convergence (in Frobenius norm)}

\item{maxiter}{maximum iteration}
}
\description{
Power method for leading eigenvectors and eigenvalues
}
