% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{println}
\alias{println}
\title{Print line}
\usage{
println(...)
}
\arguments{
\item{...}{R objects printable using cat()}
}
\description{
Print line
}
\details{
This is a wrapper function that prints R objects using cat() with
no space separation and end with a newline character
}
