% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{pseudoinverse}
\alias{pseudoinverse}
\title{Compute Moore-Penrose pseudoinverse of a real rectangular matrix}
\usage{
pseudoinverse(M, tolerance = 1e-10)
}
\arguments{
\item{M}{real matrix to be inverted}

\item{tolerance}{singular values below this level are treated as 0}
}
\value{
matrix M.inv satisfying M %*% M.inv %*% M = M and M.inv %*% M %*% M.inv = M.inv
}
\description{
Compute Moore-Penrose pseudoinverse of a real rectangular matrix
}
