% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\docType{package}
\name{putils}
\alias{putils}
\alias{putils-package}
\title{putils: A package for personal utility functions.}
\description{
The putils package provides personal utility functions.
}
\section{mathematical functsions}{

\itemize{
  \item expit
  \item logit
}
* expit
* logit
* lambertW
}

\section{matrices and vectors}{

* vector.norm
* vector.normalise
* vector.clip
* vector.soft.thresh
* vector.hard.thresh
* vector.scramble
* vector.cleanup
* matrix.GramSchmidt
* matrix.trace
* matrix.rank
* matrix.standardise
* matrix.power
* sinThetaLoss
* powerMethod
* ql
* offdiag
}

\section{random element generation}{

* random.rademacher
* random.bernoulli
* random.UnitVector
* random.OrthogonalMatrix
* random.WishartMatrix
* random.WignerMatrix
* random.psdMatrix
* random.SymmetricMatrix
}

\section{auxiliary functions}{

* printPercentage
* visualise
* snippet
* find.first
* find.last
* sf
* sf_exp
* sim.params
* show.params
* '%=%'
}

\section{statistical functions}{

* CvM.test
}

\section{string operations}{

* strlen
* strstr
* prefix
* suffix
}

\section{NA handling}{

* setNA
}

