% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{QR}
\alias{QR}
\title{QR decomposition of a matrix}
\usage{
QR(A)
}
\arguments{
\item{A}{an nxp matrix for QR decomposition}
}
\value{
a list of two matrices Q and R so that A = QR
\itemize{
  \item Q - nxp matrix with orthonormal columns with the same span as A
  \item R - a upper triangular pxp matrix with nonnegative diagonal entries
}
}
\description{
QR decomposition of a matrix
}
