% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random.GaussianMatrix}
\alias{random.GaussianMatrix}
\title{Generate a Gaussian random matrix}
\usage{
random.GaussianMatrix(nrow, ncol = nrow)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Generate a Gaussian random matrix
}
