% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random.OrthogonalMatrix}
\alias{random.OrthogonalMatrix}
\title{Generate a random matrix with orthonormal columns}
\usage{
random.OrthogonalMatrix(nrow, ncol = nrow)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns (should be no larger than nrow)}
}
\description{
Generate a random matrix with orthonormal columns
}
