% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{random.psdMatrix}
\alias{random.psdMatrix}
\title{Generate a random positive semidefinite matrix with operator norm <=1}
\usage{
random.psdMatrix(n)
}
\arguments{
\item{n}{dimension}
}
\description{
Generate a random positive semidefinite matrix with operator norm <=1
}
