% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{random.rademacher}
\alias{random.rademacher}
\title{generate n rademacher random variables}
\usage{
random.rademacher(n = 1)
}
\arguments{
\item{n}{length of random vector}
}
\description{
generate n rademacher random variables
}
