% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{random.SymmetricMatrix}
\alias{random.SymmetricMatrix}
\title{generate a random symmetric matrix}
\usage{
random.SymmetricMatrix(n)
}
\arguments{
\item{n}{dimension}
}
\description{
generate a random symmetric matrix
}
