% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{random.UnitVector}
\alias{random.UnitVector}
\title{generate a random unit vectors in R^n}
\usage{
random.UnitVector(n)
}
\arguments{
\item{n}{length of random vector}
}
\description{
generate a random unit vectors in R^n
}
