% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random.WignerMatrix}
\alias{random.WignerMatrix}
\title{Generate a random nxn Wigner matrix, where off diagonals are N(0,1) and diagonals are N(0,2)}
\usage{
random.WignerMatrix(n)
}
\arguments{
\item{n}{dimension}
}
\description{
Generate a random nxn Wigner matrix, where off diagonals are N(0,1) and diagonals are N(0,2)
}
