% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{random.WishartMatrix}
\alias{random.WishartMatrix}
\title{Generate a random nxn Wishart matrix}
\usage{
random.WishartMatrix(n)
}
\arguments{
\item{n}{dimension}
}
\description{
Generate a random nxn Wishart matrix
}
