% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{rep_along}
\alias{rep_along}
\title{Repeat a vector along another vector to achieve equal length}
\usage{
rep_along(x, y)
}
\arguments{
\item{x}{vector to be repeated}

\item{y}{vector whose length sholud be matched}
}
\description{
Repeat a vector along another vector to achieve equal length
}
