% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{setNA}
\alias{setNA}
\title{Change all NA values in v to a}
\usage{
setNA(v, a)
}
\arguments{
\item{v}{a vector}

\item{a}{target value}
}
\value{
updated vector
}
\description{
Change all NA values in v to a
}
