% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{sf}
\alias{sf}
\title{display signif nicely}
\usage{
sf(x, digits)
}
\arguments{
\item{x}{a real number}

\item{digits}{number of significant figures to keep}
}
\value{
a string
}
\description{
display signif nicely
}
\details{
keep trailing zeros, not use scientific notation
}
