% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sf_exp}
\alias{sf_exp}
\title{display signif of exponentiated number nicely}
\usage{
sf_exp(x, digits)
}
\arguments{
\item{x}{a real number}

\item{digits}{positive integer, number of significant figures}
}
\value{
a string
}
\description{
display signif of exponentiated number nicely
}
\details{
significant figure computed after subtracting 1. keep trailing zeros, not use scientific notation
}
