% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{show.params}
\alias{show.params}
\title{Show parameter values}
\usage{
show.params(...)
}
\arguments{
\item{...}{variables to be printed or a named vector of parameters}
}
\description{
Print out parameters in a vector in a nice format
}
