% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{snippet}
\alias{snippet}
\title{Show snippet of a large vector/matrix}
\usage{
snippet(A, nrow = 5, ncol = nrow)
}
\arguments{
\item{A}{a vector, matrix or array}
}
\description{
Show snippet of a large vector/matrix
}
