% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{sort.data.frame}
\alias{sort.data.frame}
\title{Sort data frame by columns}
\usage{
\method{sort}{data.frame}(x, by = 1, decreasing = FALSE)
}
\arguments{
\item{x}{a data frame}

\item{by}{vector of column indices, sort by first index first, ties are broken using second index etc.}
}
\description{
Sort data frame by columns
}
