% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sort_data_frame}
\alias{sort_data_frame}
\title{sort dataframe}
\usage{
sort_data_frame(df, sort_by, decreasing = FALSE)
}
\arguments{
\item{df}{dataframe to be sorted}

\item{sort_by}{column names to sort by, in decreasing order of priority}

\item{decreasing}{a vector of the same length as sort_by, recycled to match length if needed}
}
\description{
sort dataframe
}
