% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{standardise}
\alias{standardise}
\title{Standardise the columns of a matrix to have mean zero and length 1}
\usage{
standardise(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
a standardised matrix
}
\description{
Standardise the columns of a matrix to have mean zero and length 1
}
