% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{strstr}
\alias{strstr}
\title{Substring search}
\usage{
strstr(haystack, needle, startpos = 1, n = 1)
}
\arguments{
\item{haystack}{a string}

\item{needle}{substring to search for}

\item{startpos}{start position for search}

\item{n}{the nth occurrence}
}
\value{
an integer
}
\description{
find the position of the nth occurrence of needle in haystack, returns 0 if not found
}
