% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{stylise}
\alias{stylise}
\title{assign line types to distinct values of a vector}
\usage{
stylise(v)
}
\arguments{
\item{v}{vector to be converted to line types / categories}
}
\value{
a vector of line types of equal length with two attributes: 'lty' giving a vector of all line types used; 'uniq': vector of distinct values in v
}
\description{
assign line types to distinct values of a vector
}
