% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sweep_arithmetic}
\alias{sweep_arithmetic}
\alias{\%_+_\%}
\alias{\%_-_\%}
\alias{\%_*_\%}
\alias{\%_/_\%}
\title{Binary operators, add/subtract/multiply/divide a vector to a matrix row by
row. i.e. each row of the matrix is added/subtracted/etc by the same vector}
\usage{
x \%_+_\% y

x \%_-_\% y

x \%_*_\% y

x \%_/_\% y
}
\arguments{
\item{x}{matrix or vector}

\item{y}{matrix or vector (one of x and y needs to be a vector of length
equal to the number of columns of the other)}
}
\description{
Binary operators, add/subtract/multiply/divide a vector to a matrix row by
row. i.e. each row of the matrix is added/subtracted/etc by the same vector
}
