% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{vector.cleanup}
\alias{vector.cleanup}
\title{Clean up floating point epsilon terms}
\usage{
vector.cleanup(v, tolerance = 1e-14)
}
\arguments{
\item{v}{a vector}

\item{tolerance}{level below which we set entries to 0}
}
\value{
a vector of the same length
}
\description{
Clean up floating point epsilon terms
}
