% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{vector.scramble}
\alias{vector.scramble}
\title{Randomly shuffle a vector}
\usage{
vector.scramble(v)
}
\arguments{
\item{v}{a vector}
}
\value{
a shuffled vector of the same length
}
\description{
Randomly shuffle a vector
}
