% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{visualise}
\alias{visualise}
\title{Visualise a matrix X}
\usage{
visualise(
  X,
  aspect.ratio = c("automatic", "actual"),
  axes = FALSE,
  frame.plot = FALSE
)
}
\arguments{
\item{X}{a matrix}

\item{aspect.ratio}{if automatic, it will be calculated automatically to fit screen, otherwise, the actual dimension of the matrix will be used.}

\item{axes}{whether to display axes}

\item{frame.plot}{whether to draw a frame}
}
\value{
a color plot of matrix value magnitude
}
\description{
Visualise a matrix X
}
