% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{\%=\%}
\alias{\%=\%}
\title{Multiple assignment}
\usage{
l \%=\% r
}
\arguments{
\item{l}{left side list, enclosed by the \code{bunch} function}

\item{r}{right side list}
}
\description{
assign multiple items in a list on RHS to multiple items in a list on LHS
}
\details{

}
