% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random.bernoulli}
\alias{random.bernoulli}
\title{Generate n random Bernoulli variables}
\usage{
random.bernoulli(n = 1, p = 0.5)
}
\arguments{
\item{n}{length of random vector}

\item{p}{probability (can be a vector)}
}
\description{
Generate n random Bernoulli variables
}
