% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{random.OrthogonalMatrix}
\alias{random.OrthogonalMatrix}
\title{Generate a random nxn orthogonal matrix}
\usage{
random.OrthogonalMatrix(n)
}
\arguments{
\item{n}{dimension}
}
\description{
Generate a random nxn orthogonal matrix
}
