% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random.SymmetricMatrix}
\alias{random.SymmetricMatrix}
\title{Generate a random symmetric matrix}
\usage{
random.SymmetricMatrix(n)
}
\arguments{
\item{n}{dimension}
}
\description{
Generate a random symmetric matrix
}
