% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{sinThetaLoss}
\alias{sinThetaLoss}
\title{Sine angle loss}
\usage{
sinThetaLoss(U, V)
}
\arguments{
\item{U}{an orthonormal matrix}

\item{V}{an orthonormal matrix}
}
\value{
sine angle loss between Vhat and V (Frobenius norm)
}
\description{
Sine angle loss
}
