% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R, R/linalg.R
\name{sort.data.frame}
\alias{sort.data.frame}
\title{sort dataframe}
\usage{
\method{sort}{data.frame}(x, by = 1, decreasing = FALSE)

\method{sort}{data.frame}(x, by = 1, decreasing = FALSE)
}
\arguments{
\item{x}{a data frame}

\item{by}{vector of column indices, sort by first index first, ties are broken using second index etc.}

\item{decreasing}{a vector of the same length as sort_by, recycled to match length if needed}

\item{df}{dataframe to be sorted}

\item{sort_by}{column names to sort by, in decreasing order of priority}
}
\description{
sort dataframe

Sort data frame by columns
}
